changed it to use Integer not int everywhere
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@6717 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/test/java/org/drools/examples/manners/Count.java
@@ -16,21 +16,21 @@
 package org.drools.examples.manners;
 
 public class Count {
-    private int value;
+    private Integer value;
     
     public Count() {
     }
 
-    public Count(final int value) {
+    public Count(final Integer value) {
         super();
         this.value = value;
     }
 
-    public int getValue() {
+    public Integer getValue() {
         return this.value;
     }
 
-    public void setValue(final int value) {
+    public void setValue(final Integer value) {
         this.value = value;
     }
 
@@ -51,7 +51,7 @@ public class Count {
     }
 
     public int hashCode() {
-        return this.value;
+        return this.value.intValue();
     }
 
 }
